// api/src/menus/interfaces/menu.interface.ts
export interface Menu {
      Name: string;
      Category: string;      
      About: string;
      Price: string;
      Available: string;
      Image: string;
      RestaurantId: any;
      CategoryId : any;
      IsSpecial: boolean,
      IsDealOfTheDay: boolean,
      Toppings: any,
      Counts: Number,
}